Faster libtool

+++ tools/objc/libtool.sh
@@ -34,6 +34,16 @@ fi
 
 WRAPPER="${MY_LOCATION}/xcrunwrapper.sh"
 
+# Ensure 0 timestamping for hermetic results.
+export ZERO_AR_DATE=1
+
+if `dirname "$0"`/libtool_fast_path "$@"; then
+  "${WRAPPER}" libtool "$@"
+  exit
+fi
+
+echo "Duplicate .o basenames detected, taking slow path"
+
 TEMPDIR="$(mktemp -d "${TMPDIR:-/tmp}/libtool.XXXXXXXX")"
 trap "rm -rf \"${TEMPDIR}\"" EXIT
 
@@ -111,11 +121,4 @@ while [[ $# -gt 0 ]]; do
    esac
 done
 
-# Ensure 0 timestamping for hermetic results.
-export ZERO_AR_DATE=1
-
-"${WRAPPER}" libtool "${ARGS[@]}"
-
-# Prevents a pre-Xcode-8 bug in which passing zero-date archive files to ld
-# would cause ld to error.
-touch "$OUTPUTFILE"
+"${WRAPPER}" libtool "$@" "${ARGS[@]}"
